Drop test view when done with it.
The view just added by commit dae2543f4 decompiles differently
in v15 than HEAD (presumably as a consequence of dfb35441d).
That causes failures in cross-version upgrade testing.

We could teach AdjustUpgrade.pm to compensate for that, but it
seems less painful to just drop the view after we're done with it.

Per buildfarm.